src/main/requirements.txt
attrs==20.3.0
coverage==5.4
emitter @ file:///home/segu_dev/workspace/python_dev/streamer/src/main/venv/emitter-1.0.0.2-py3-none-any.whl
iniconfig==1.1.1
kafka-python==2.0.2
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-logstash==0.4.6
toml==0.10.2
